change: apply user id type

+++ src/main/kotlin/com/assignment/caffe/adapter/in/web/SignController.kt
@@ -58,6 +58,6 @@ class SignController(
         @RequestHeader("Authorization") token: String,
         principal: Principal,
     ) {
-        signUseCase.signOut(principal.name.toInt(), token)
+        signUseCase.signOut(principal.name, token)
     }
 }
